Distinguish between SwaggerMetadataConverter and SwaggerMetadataProvider
