+++ src/Nancy.Swagger.Annotations/Nancy.Swagger.Annotations.csproj
@@ -61,7 +61,7 @@
     <Compile Include="Extensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RouteId.cs" />
-    <Compile Include="SwaggerAnnotationsConverter.cs" />
+    <Compile Include="SwaggerAnnotationsProvider.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Nancy.Swagger\Nancy.Swagger.csproj">
