+++ src/Nancy.Swagger/Services/ISwaggerMetadataProvider.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Nancy.Swagger.Services
+{
+    public interface ISwaggerMetadataProvider
+    {
+        IList<SwaggerModelData> RetrieveSwaggerModelData();
+
+        IList<SwaggerRouteData> RetrieveSwaggerRouteData();
+    }
+}
