+++ src/Nancy.Swagger/SwaggerRegistrations.cs
@@ -8,6 +8,7 @@ namespace Nancy.Swagger
     {
         public SwaggerRegistrations()
         {
+            RegisterWithDefault<ISwaggerMetadataProvider>(typeof(DefaultSwaggerMetadataProvider));
             RegisterWithDefault<ISwaggerMetadataConverter>(typeof(DefaultSwaggerMetadataConverter));
             RegisterWithDefault<ISwaggerModelCatalog>(typeof(DefaultSwaggerModelCatalog));
             RegisterAll<ISwaggerModelDataProvider>();
