+++ test/Nancy.Swagger.Annotations.Tests/Nancy.Swagger.Annotations.Tests.csproj
@@ -61,7 +61,7 @@
     <Compile Include="..\..\etc\CommonAssemblyInfo.cs">
       <Link>Properties\CommonAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="SwaggerAnnotationsConverterTests.cs" />
+    <Compile Include="SwaggerAnnotationsProviderTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestModel.cs" />
     <Compile Include="TestModel.InOtherNamespace.cs" />
