+++ test/Nancy.Swagger.Annotations.Tests/SwaggerAnnotationsProviderTests.Get_ApiDocsPath_ReturnsApiDeclaration.approved.json
