+++ test/Nancy.Swagger.Annotations.Tests/SwaggerAnnotationsProviderTests.Get_ApiDocsRootpath_ReturnsResourceListing.approved.json
