+++ test/Nancy.Swagger.Annotations.Tests/SwaggerAnnotationsProviderTests.Get_TestModulePath_ReturnsApiDeclaration.approved.json
